refactor(capabilities): hoist shared content-gen HTTP plumbing to clear qodana dups
Extracts the byte-identical HTTP scaffolding the two content-gen
adapters carried — the configured ureq agent, the blocking
request-run block, the status-prefix parse, and the body-snippet trim —
into contentgen::shared, and the cap unit tests' decode_reply helper
into test_chassis::decode_session_reply. Also drops the module-prefix
on two contentgen doc-links. Clears the 8 Qodana findings (6 duplicated
fragments + 2 unnecessary path prefixes) on PR 1040; the per-provider
error taxonomy stays distinct.
